+++ lib/events/offramp_sale_created.dart
@@ -1,0 +1,5 @@
+part of '../widget_event.dart';
+
+final class OfframpSaleCreated extends WidgetEvent {
+  const OfframpSaleCreated();
+}