+++ lib/events/purchase_created.dart
@@ -1,0 +1,5 @@
+part of '../widget_event.dart';
+
+final class PurchaseCreated extends WidgetEvent {
+  const PurchaseCreated();
+}